reorganizes project in preperation for 'Interaction' landing
